+++ android/src/main/kotlin/com/prateekmedia/muxer/MuxerPlugin.kt
@@ -1,4 +1,4 @@
-package com.prateekmedia.mp4parser
+package com.prateekmedia.muxer
 
 import androidx.annotation.NonNull
 
@@ -17,8 +17,8 @@ import java.io.FileInputStream
 import java.io.FileOutputStream
 import java.io.RandomAccessFile
 
-/** Mp4parserPlugin */
-class Mp4parserPlugin: FlutterPlugin, MethodCallHandler {
+/** MuxerPlugin */
+class MuxerPlugin: FlutterPlugin, MethodCallHandler {
   /// The MethodChannel that will the communication between Flutter and native Android
   ///
   /// This local reference serves to register the plugin with the Flutter Engine and unregister it
@@ -26,7 +26,7 @@ class Mp4parserPlugin: FlutterPlugin, MethodCallHandler {
   private lateinit var channel : MethodChannel
 
   override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
-    channel = MethodChannel(flutterPluginBinding.binaryMessenger, "mp4parser")
+    channel = MethodChannel(flutterPluginBinding.binaryMessenger, "muxer")
     channel.setMethodCallHandler(this)
   }