+++ example/lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'dart:async';
 
 import 'package:flutter/services.dart';
-import 'package:mp4parser/mp4parser.dart';
+import 'package:muxer/muxer.dart';
 
 void main() {
   runApp(const MyApp());
@@ -31,12 +31,15 @@ class _MyAppState extends State<MyApp> {
     // We also handle the message potentially returning null.
     try {
       platformVersion =
-          await Mp4parser.muxAudioVideo(videoPath: '', audioPath: '') ??
-              'Mux failed';
+          await Muxer.muxAudioVideo(audioPath: '', videoPath: '') ??
+              'Unknown platform version';
     } on PlatformException {
-      platformVersion = 'Failed to mux audio video.';
+      platformVersion = 'Failed to get platform version.';
     }
 
+    // If the widget was removed from the tree while the asynchronous platform
+    // message was in flight, we want to discard the reply rather than calling
+    // setState to update our non-existent appearance.
     if (!mounted) return;
 
     setState(() {