+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.1.0 - Let there be light
 
-* Initial release of mp4parser for dart
+* Initial release of muxer for flutter